+++ src/core/injections/semiblind/techniques/file_based/fb_injector.py
@@ -326,7 +326,7 @@ def injection_results(url, OUTPUT_TEXTFILE, timesec):
   headers.do_check(request)
   headers.check_http_traffic(request)
   # Check if defined any HTTP Proxy (--proxy option).
-  if menu.options.proxy:
+  if menu.options.proxy or menu.options.ignore_proxy: 
     response = proxy.use_proxy(request)
   # Check if defined Tor (--tor option).
   elif menu.options.tor: